Moving notificiation into it's own file
