+++ src/scripting/Elsa.Scripting.JavaScript/Messages/EvaluatingJavaScriptExpression.cs
@@ -15,15 +15,4 @@ namespace Elsa.Scripting.JavaScript.Messages
         public Engine Engine { get; }
         public ActivityExecutionContext ActivityExecutionContext { get; }
     }
-
-    public class ConfigureJavaScriptOptions : INotification
-    {
-        public ConfigureJavaScriptOptions(Jint.Options options, ActivityExecutionContext context)
-        {
-            JintOptions = options;
-            ActivityExecutionContext = context;
-        }
-        public Jint.Options JintOptions { get; set; }
-        public ActivityExecutionContext ActivityExecutionContext { get; }
-    }
 }
